Refactor recommendation services into unified module
