+++ sidetrack/worker/jobs.py
@@ -12,7 +12,7 @@ from sidetrack.api.main import aggregate_weeks as aggregate_weeks_service
 from sidetrack.services.listens import get_listen_service
 from sidetrack.common.models import Feature, Track
 from sidetrack.services.datasync import sync_user as datasync_sync_user
-from sidetrack.services.insights import compute_weekly_insights
+from sidetrack.services.recommendation import compute_weekly_insights
 from sidetrack.services.listenbrainz import ListenBrainzClient
 from sidetrack.services.musicbrainz import MusicBrainzService
